chore: refactor ABI structs and consolidate modules
Moved all Solidity ABI struct definitions to a new `abi` module for better organization and maintainability. Updated references to these structures across the codebase, ensuring consistency and removing redundancies.
