+++ src/utils/v4_planner.rs
@@ -1,6 +1,6 @@
-use crate::prelude::{encode_route_to_path, Error, PathKey, Trade};
+use crate::prelude::{encode_route_to_path, Error, Trade, *};
 use alloy_primitives::{Bytes, U256};
-use alloy_sol_types::{sol, SolValue};
+use alloy_sol_types::SolValue;
 use uniswap_sdk_core::prelude::*;
 use uniswap_v3_sdk::prelude::*;
 
@@ -72,160 +72,6 @@ impl Actions {
             Self::SWEEP(params) => params.abi_encode(),
         }
         .into()
-    }
-}
-
-sol! {
-    #[derive(Debug, Default, PartialEq)]
-    struct PoolKeyStruct {
-        address currency0;
-        address currency1;
-        uint24 fee;
-        int24 tickSpacing;
-        address hooks;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct IncreaseLiquidityParams {
-        uint256 tokenId;
-        uint256 liquidity;
-        uint128 amount0Max;
-        uint128 amount1Max;
-        bytes hookData;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct DecreaseLiquidityParams {
-        uint256 tokenId;
-        uint256 liquidity;
-        uint128 amount0Min;
-        uint128 amount1Min;
-        bytes hookData;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct MintPositionParams {
-        PoolKeyStruct poolKey;
-        int24 tickLower;
-        int24 tickUpper;
-        uint256 liquidity;
-        uint128 amount0Max;
-        uint128 amount1Max;
-        address owner;
-        bytes hookData;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct BurnPositionParams {
-        uint256 tokenId;
-        uint128 amount0Min;
-        uint128 amount1Min;
-        bytes hookData;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct SwapExactInSingleParams {
-        PoolKeyStruct poolKey;
-        bool zeroForOne;
-        uint128 amountIn;
-        uint128 amountOutMinimum;
-        uint160 sqrtPriceLimitX96;
-        bytes hookData;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct SwapExactInParams {
-        address currencyIn;
-        PathKey[] path;
-        uint128 amountIn;
-        uint128 amountOutMinimum;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct SwapExactOutSingleParams {
-        PoolKeyStruct poolKey;
-        bool zeroForOne;
-        uint128 amountOut;
-        uint128 amountInMaximum;
-        uint160 sqrtPriceLimitX96;
-        bytes hookData;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct SwapExactOutParams {
-        address currencyOut;
-        PathKey[] path;
-        uint128 amountOut;
-        uint128 amountInMaximum;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct SettleParams {
-        address currency;
-        uint256 amount;
-        bool payerIsUser;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct SettleAllParams {
-        address currency;
-        uint256 maxAmount;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct SettlePairParams {
-        address currency0;
-        address currency1;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct TakeParams {
-        address currency;
-        address recipient;
-        uint256 amount;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct TakeAllParams {
-        address currency;
-        uint256 minAmount;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct TakePortionParams {
-        address currency;
-        address recipient;
-        uint256 bips;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct TakePairParams {
-        address currency0;
-        address currency1;
-        address recipient;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct SettleTakePairParams {
-        address settleCurrency;
-        address takeCurrency;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct CloseCurrencyParams {
-        address currency;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct SweepParams {
-        address currency;
-        address recipient;
-    }
-
-    #[derive(Debug, Default, PartialEq)]
-    struct ActionsParams {
-        bytes actions;
-        bytes[] params;
     }
 }
